fix(a11y): reuse skip-to-content link
